src safe - add displaying video picture and play button - todo: navigate to video player
