+++ app/src/main/java/dev/leonlatsch/photok/ui/imageviewer/ImageViewerViewModel.kt
@@ -14,7 +14,7 @@
  *   limitations under the License.
  */
 
-package dev.leonlatsch.photok.ui.viewphoto
+package dev.leonlatsch.photok.ui.imageviewer
 
 import android.app.Application
 import androidx.databinding.Bindable
